templates/requirements.txt

flask>=3.0.0
python-dotenv>=1.0.1
google-genai>=1.0.0
google-generativeai>=0.8.3
markdown>=3.6
einops>=0.7.0
torch>=2.1.0
